refactor: components 폴더내에 있는 header 관련 컴포넌트들 components/header로 이동
- BattleHeader.tsx / StatusCard.tsx 모두 Header 관련 기능으로 /header 폴더로 이동 했습니다.
