+++ frontend/src/pages/battlePage/components/header/BattleHeader.tsx
