+++ frontend/src/pages/battlePage/components/header/StatusCard.tsx
